mabasa_backend/requirements.txt
Django==5.0.6
 djangorestframework==3.15.2
 djangorestframework-simplejwt==5.3.1
 django-cors-headers==4.4.0
